server/requirements.txt
fastapi==0.109.1
uvicorn==0.27.1
python-dotenv==1.0.0
aiohttp==3.9.1
pydantic==2.5.3
python-multipart==0.0.6
websockets==12.0
openai==1.12.0 